+++ app/src/androidTest/java/campus/tech/kakao/map/uiTest/MapActivitiyTest.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.uiTest
 
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
@@ -7,6 +7,8 @@ import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import campus.tech.kakao.map.R
+import campus.tech.kakao.map.ui.MapActivity
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
