+++ app/src/androidTest/java/campus/tech/kakao/map/unitTest/MainActivityUnitTest.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.unitTest
 
 import android.view.View
 import android.widget.EditText
@@ -7,6 +7,8 @@ import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.RecyclerView
 import androidx.test.core.app.ActivityScenario
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import campus.tech.kakao.map.R
+import campus.tech.kakao.map.ui.MainActivity
 import junit.framework.TestCase.assertEquals
 import junit.framework.TestCase.assertNotNull
 import junit.framework.TestCase.assertTrue
