+++ app/src/main/AndroidManifest.xml
@@ -14,11 +14,11 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Map"
         tools:targetApi="31"
-        android:name=".AppClass"
+        android:name=".application.AppClass"
         android:networkSecurityConfig="@xml/network_security_config">
 
         <activity
-            android:name=".MapActivity"
+            android:name=".ui.MapActivity"
             android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
@@ -27,7 +27,7 @@
         </activity>
 
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="false" />
 
     </application>
