+++ app/src/main/java/campus/tech/kakao/map/database/MapContract.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.database
 
 object MapContract
 {
