+++ app/src/main/java/campus/tech/kakao/map/network/KakaoMapRetrofitService.kt
@@ -1,5 +1,6 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.network
 
+import campus.tech.kakao.map.model.KakaoMapProductResponse
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Header
