+++ app/src/main/java/campus/tech/kakao/map/ui/KeywordAdapter.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.ui
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,6 +6,7 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import campus.tech.kakao.map.R
 
 class KeywordAdapter(private val listener: OnKeywordRemoveListener) : RecyclerView.Adapter<KeywordAdapter.ViewHolder>() {
