+++ app/src/main/res/layout/activity_map.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MapActivity">
+    tools:context=".ui.MapActivity">
 
     <LinearLayout
         android:id="@+id/searchLayout"
